Remove the id-receiving-slots
They are now useless as the id is given by the Manager rather than the
Galago server, which makes KNotification quite depend on it. This way
the ids are fully independent and the mapping between Galago/DBus ids
and KNotification ids is done only in the NotifyByPopup plugin, where it
belongs.