Port NotifyByPopup to KF5/Qt5
This is rather a major rewrite of this plugin. Main changes:

 * all methods dealing with the Galago server are renamed as such
   to indicate more clearly what they do
 * async dbus calls are used where possible
 * code is simplified and cleaned up; lots of comments added

NotifyByPopup keeps the KNotification<-->Galago mapping internally and
is the only part that deals with DBus now. The flow of putting a
notification on screen is now this:

KNotification::sendEvent() -> KNotificationManager::notify() -> NotifyByPopup::notify() -> DBus -> Galago server

whereas before it was something like this:

KNotification::sendEvent() -> KNotificationManager::notify() -> org::kde::KNotify::notify() -> DBus -> KNotify::_the_dbus_handler -> NotifyByPopup::notify() -> DBus -> Galago server

It also keeps the fallback to Growl or KPassivePopup in case the Galago
server is not available