Fix crash accessing dangling pointer in NotifyByPopup
sendNotificationToGalagoServer creates a new QDBusPendingCallWatcher for
a sent DBus message. It puts the original KNotification as a proprty on
the QObject.

If the local KNotification object got destroyed whilst that DBus
operation was happening this meant we would read a dangling pointer.

By setting the parent of the call watcher to the notification we can
make sure we don't process the call finishing if the notification no
longer exists.

BUG: 340845
REVIEW: 121786