Get rid of KDBusMenuExporter and use DBusMenuExporter directly
KDBusMenuExporter seems to exist only to "provide access to KDE icon
names" but the code in DBusMenuExporter is virtually the same, so
there's no use for KDBusMenuExporter now.

This had a use in kde4 world where it used KIcon, but QIcon now uses
KIconEngine on our platform anyway (and also in pre-qt4.7 before
QIcon::name() was a thing).

REVIEW: 121864