Added an event() version that takes no icon and will use a default one
Had to remove a default empty QPixmap parameter from a different event()
header, that would make a call to one of them ambiguous. This means that
when recompiling an app that used event() with StandardEvent and didn't
specify an icon, it will now pick the version that uses standard icons
instead of the one that shows no icon at all. (Which seems ok to me).

REVIEW: 122913