Introduce LoopSound flag
Some notifications might want to loop the sound, eg. calling apps
playing the ringing sound which is not wanted to play once. Currently
the sound is looped for all persistent notifications, but that's not
ideal as not all persistent notifications with sound want/need sound
looping. This new LoopSound flag solves that.

BUG: 346148
REVIEW: 123360
CHANGELOG: Introduce LoopSound flag allowing notifications to play sound
in a loop if they need it