Don't modify the KNotification object from NotifyByPopup plugin
The KNotification object itself should never be modified from the
plugins; the plugins should do only local checking of the properties.
Modifying the KNotification object itself can break other plugins.

This together with the previous commit also fix KNotification usage
together with Ubuntu's NotifyOSD which could make it look like stuck
while it was actually in an update loop (the server does not support
actions so it was unset on KNotification which triggered an update which
would reshow the popup which would unset the actions on KNotification
which would trigger an update and so on).

CHANGELOG: Fix KNotifications when used with Ubuntu's NotifyOSD
REVIEW: 123540
BUG: 345973