Remove KService and KIconThemes usage from KNotifications
This patch reduces the dependencies of KNotifications framework and
effectively makes it a tier 2 framework.

KService is used only for locating additional notification plugins and
to my knowledge, there are none such plugins currently existing, at
least not in all around KDE plus the class for the plugins wasn't
exported until about two months ago, so this should be safe without
a legacy support.

KIconThemes is used only to get "KIconLoader::Small" icon pixmap for
notifications using KPassivePopup. After some playing around it turns
out that it puts the icon into the KPassivePopup title and makes it as
big as the text. So I've made the icon size to be the same as the text
height. So this keeps things visually the same + still DPI aware,
though I believe the KPassivePopup should receive a complete visual
overhaul anyway.

Additionally, KCodecs dependency has again one single usage for decoding
html entities to QChars within QXmlStreamReader parser, so eventually
could also be removed/replaced with QTextDocument::toPlainText() which
seems to do the same job as QXmlStreamReader+KCodecs.

REVIEW: 124281
CHANGELOG: Reduce dependencies and move to Tier 2