Add a hint to skip notifications grouping
Plasma currently groups notifications that have the same
notification title or if they are on a "blacklist" (like
media player apps that popup a notification for each song
change), but sometimes the grouping needs to be skipped.
This is for example when user removes two widgets from his
desktop and the "Undo" notification appears, these cannot
be grouped.

This is currently handled rather by a hack in the notifications
server, so introducing this new flag should allow for a clean
and proper solution.

REVIEW: 126459
CCBUG: 356653