Use application name instead of pid when creating SNI dbus service
Summary:
This is required when using SNI in sandbox (e.g. in flatpak) where the pid starts from 1,
usually it is 2 (the first one is bwrap) and this can theoretically cause collision with
already registered services. Also with using app name instead of pid we know in advance
what the service name will look like and we can give the sandboxed app access to it, which
is not possible when using pid as the name is not known until the app is started.

Reviewers: apol, mart

Reviewed By: mart

Subscribers: mart, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5783