Don't block starting notification service
Summary:
We don't need to manually start the DBus service.
It blocks the calling app, and dbusServiceExists means that we will
always end up going the DBus route over a popup anyway, so it won't
do anything useful.

The service (in the plasma case plasma-wait-for-name) will be started
automatically when we actually send the notification.

Also fix d-dbusServiceExists being cleared to false when the first
service owner exits.

BUG: 382444

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D6963