Remove check for SNI hosts when chosing whether to use legacy mode
Summary:
Previous code (since very first commit) checked for both having a valid
KSNI watcher as well as making sure there was a host available in order
to choose which mode to use.

This made some sense when Plasma 4 still had proper xembed support and
one could turn SNIs off.

The current code here is broken. On first creation we only check for a
watcher - and only ends up checking for the host if it changes at
runtime.

Given it's not working properly, and the use case has changed it seems
we can safely remove this code.

This removes a blocking DBus call on the property.

Reviewers: graesslin

Reviewed By: graesslin

Subscribers: graesslin, #frameworks
BUG: 385867

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8354