Notification portal: support pixmaps in notifications
Summary:
From portal documentation:
The portal only accepts serialized GThemedIcon and GBytesIcons. Both of
these have the form (sv). For themed icons, the string is "themed",
and the value is an array of strings containing the icon names.
For bytes icons, the string is "bytes", and the value is a bytestring
containing the icon data in png, jpeg or svg form. For historical
reasons, it is also possible to send a simple string for themed
icons with a single icon name.

Reviewers: broulik, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17188