Build without D-Bus on Android
Summary:
This is probably the most invasive part of the Android support,
the remainder is implementing the native Android backend.

This will mean that KF5::Notifications will miss some API on Android
that depends on D-Bus, but with D-Bus not only being non-functional but
usually not even available as a stub I don't really see how to avoid that.

Reviewers: dhaumann

Reviewed By: dhaumann

Subscribers: bcooksley, apol, dhaumann, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17817