Build the Java side with Gradle, as AAR instead of JAR
Summary:
Most importantly this allows us to use the AndroidX support library, to
properly support a wide range of different Android versions (we link to
that for demonstration purposes only atm, it's not actually used yet).

This is closer to the Android way of doing things than the previous raw
approach, which should be more sustainable. The obvious dirty things this
does like copying the build tool into the project or downloading tons of
stuff during the build seem to be "normal".

For this to be consumable by application, a tiny change to their
build.gradle files is needed: https://codereview.qt-project.org/#/c/249179/

Reviewers: apol

Reviewed By: apol

Subscribers: apol, nicolasfella, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17986