Move the Android API level checks to runtime
Summary:
This doesn't make notifications on API level < 23 work yet, but it allows
us to build against lower API levels and have something that still works
when run on higher API levels.

Tested on API level 21 and 27.

This should also enable building the Android support on the CI.

Reviewers: apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D18253