Deprecate raiseWidget
Summary: It is unused. Furthermore, it relies on QtWidgets API which we try to get rid of in KNotifications. I assume the intention of this method was to raise the window when the notification is clicked, but this should be done by settng a default action and using appropriate QWindow API.

Reviewers: #frameworks, broulik, apol

Reviewed By: apol

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26594