Don't use KWindowSystem on Android
Summary:
The usage of KWindowSystem is not relevant on Android. This saves us a dependency there.

Ideally we wouldn't build KPassivePopup at all there, but unfortunately it is public API. One could argue that such a break for non-functional API on Android could be okay.

Depends on D26513

Test Plan: Builds on Android and Linux

Reviewers: #frameworks, vkrause, apol, dfaure

Reviewed By: vkrause, apol, dfaure

Subscribers: dfaure, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26532