Adapt action numbering in Android to work like in KNotifications
Summary:
KNotifications assigns the ID 0 to the default action on a notification (the
one when the notification itself is selected). Additional actions (buttons on
the notification) are counted starting with 1.

In the Android backend however, the default action has ID 0 and the additional
actions are counted starting with 0 as well. This has been fixed to work as
described above.

Reviewers: #frameworks, nicolasfella, vkrause

Reviewed By: nicolasfella

Subscribers: kde-frameworks-devel, jbbgameich

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27740