Don't use notifybysnore.h on MSYS2
Summary:
SnoreToast fails to build on MSYS2 due to missing
which apparently is not available for this compiler.

This patch changes dependency on SnoreToast from required to optional.
As a result, it allows to actually build KNotifications on MSYS2.

Test Plan: Built application on Windows, which links to KNotifications.

Reviewers: broulik, brute4s99

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29258