Handle multi-line and rich-text notifications on Android
Summary:
By default only a single line of the notification message is shown,
for making longer messages readable we need to use the BigTextStyle
to get expandable multi-line messages. In the single-line case this
behaves as before.

According to the documentation BigTextStyle also should enable rich
text support, however that didn't seem to work here, so for now we
strip out rich text to not show raw HTML tags.

Reviewers: nicolasfella

Reviewed By: nicolasfella

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29323