Remove galago from method/variable naming
Summary: With D29021 the fallback notification system is gone and with it the need for differentating naming. Remove galago from the naming since it's outdated and might confuse readers unfamiliar with the code and its history. Also inline a now not needed function.

Test Plan:
Still get notifications, action invoked signal is correctly sent

Depends on D29021

Reviewers: #frameworks, broulik

Reviewed By: broulik

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29336