Implement support for notification urgency on Android
Summary:
While the notification levels map nicely, the behavior on Android with
API level 26 or higher is slightly different from what one might expect:
The urgency is configured on the notification channel the first time it
is created, and then persisted together with the user's notification
settings. That means that changes to the notification level after this
has been used for the first time have no effect.

Reviewers: nicolasfella

Reviewed By: nicolasfella

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29342