Fix removing pending notification from queue
If a notification is in the pending queue it is not yet in the
notifications map. When the notification is then closed we will hit the early return because the id wasn't found.

We need to still remove it from the queue so move the relevant block up.

If the notification is not removed from the queue it will stay as a
dangling pointer and blow up when the capabilities are queried and we
reemit the notification.

BUG: 423757