New API for notification actions
Currently actions are specified using a string list of names. When activating an action a signal like action1Activated() or activated(int index) is emitted.

This API is awkward to use and makes it hard to extend actions with further data, like icons.

Instead introduce a new KNotificationAction class that encapsulates the action and has an activated signal.

This matches what we already do for KNotificationReplyAction