Support the xdg sound spec via libcanberra
Set the sound name as an EVENT_ID property to canberra,
instead of the previous filename-based implementation.

The Url lookup to set the MEDIA_FILENAME is kept as a
fallback for non-ported configs or user selected files.
The library plays this file if the sound is not found
in the sound theme.

The sound theme is read from kdeglobals (group "Sounds",
key "Theme"), defaulting to "oxygen".