Re-implement "Run a Command" (NotifyExecute)
This is a useful feature to to run a command when an event triggers.  It
can be used to do some Text to Speech output for Calendar events. This
is really helpful if you have full screen applications which hide
notifications. You can also use it to start conferencing software prior
to a meeting.

CCBUG: 481069