Set large notification icon on Android API level >=23
Above this API level, the small icon was set to the notification's
bitmap which seems wrong. This doesn't display in the notification
drawer (presumably because it's a bitmap) and doesn't make sense
anyway because it replaces the application's icon in the status
bar.

By setting both the large and small icon, we can now properly
display the notification's desired icon along with the
application's icon. But since this large icon API is gated behind,
API 23+ I kept the small icon fallback for those older devices.